Move a bunch of RPC stuff to Actuator
This doesn't actually improve the situation.
